# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Commerce API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  stock: number;
  imageUrl: string;
  featured: boolean;
  /** @nullable */
  categoryId: number | null;
  category?: Category;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  slug: string;
  description: string;
  price: number;
  comparePrice?: number;
  stock: number;
  imageUrl: string;
  featured?: boolean;
  categoryId?: number;
}

export interface ProductUpdate {
  name?: string;
  slug?: string;
  description?: string;
  price?: number;
  comparePrice?: number;
  stock?: number;
  imageUrl?: string;
  featured?: boolean;
  categoryId?: number;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface CartItem {
  productId: number;
  quantity: number;
  product: Product;
}

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  productImageUrl: string;
  quantity: number;
  price: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  userId: number;
  status: OrderStatus;
  total: number;
  shippingAddress: string;
  items: OrderItem[];
  user?: User;
  createdAt: string;
  updatedAt?: string;
}

export interface OrderInput {
  shippingAddress: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface StoreSummary {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  totalUsers: number;
  pendingOrders: number;
  revenueThisMonth: number;
}

export interface TopProduct {
  productId: number;
  productName: string;
  productImageUrl: string;
  totalSold: number;
  revenue: number;
}

export type ListProductsParams = {
categoryId?: number;
search?: string;
minPrice?: number;
maxPrice?: number;
inStock?: boolean;
featured?: boolean;
page?: number;
limit?: number;
};

export type ListOrdersParams = {
status?: string;
page?: number;
limit?: number;
};