# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Commerce API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface CartItem {
  productId: number;
  quantity: number;
  product: Product;
}