# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Commerce API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  stock: number;
  imageUrl: string;
  featured: boolean;
  /** @nullable */
  categoryId: number | null;
  category?: Category;
  createdAt: Date;
}